Provide a filter for validity times in the streaming method
Some weather forecast sources provide data that look weeks ahead.  For
use-cases that require just the forecasts for a shorter period like 24
hours ahead, it might be overkill to stream all available data, only
to be discarded on the client.

This commit add a request parameter for filtering validity times with,
in the streaming method.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>